improve c2 deployment
